requirements.txt
streamlit==1.29.0
langchain==0.0.350
requests==2.31.0
python-dotenv==1.0.0
typing-extensions==4.8.0
pydantic==2.5.0